requirements.txt
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
paramiko==3.4.0
pillow==10.3.0
ply==3.11
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
pyparsing==3.1.2
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
